% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdata_statename_utilities.r
\name{getstfips}
\alias{getstfips}
\title{Get state fips code from state abbreviation}
\usage{
getstfips(st)
}
\arguments{
\item{st}{The state abbreviation.}
}
\description{
\code{getstfips} get state name from state fips code
}
\details{
get state fips code from state abbreviation
}
\examples{
getstfips("CO")
}
\keyword{getstfips}
