% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statename_utilities.r
\name{getstname}
\alias{getstname}
\title{Get state name from state abbreviation}
\usage{
getstname(st)
}
\arguments{
\item{st}{The state abbreviation, character.}
}
\description{
\code{getstname} get state name from state abbreviation
}
\details{
get state name from state abbreviation
}
\examples{
getstname("CO")
}
\keyword{getstname}
