% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgtax.a.r
\docType{data}
\name{sgtax.a}
\alias{sgtax.a}
\title{State government taxes, annual.}
\format{
Data frame with 1 row per state per year per variable
\describe{
\item{stabbr}{State postal abbreviation, character}
\item{year}{State fiscal year, numeric}
\item{ic}{Census Bureau item code (NA if an aggregate value)}
\item{vname}{Variable name}
\item{variable}{Variable description}
\item{value}{State tax revenue, $ thousands, numeric}
}
}
\source{
Bureau of the Census, historical database (1902-2009) and state tax files (2010-2015). See GitHub source code for details.
}
\usage{
sgtax.a
}
\description{
State government taxes, annual.
}
\examples{
  sgtax.a
}
\keyword{datasets}
