% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spi.a.r
\docType{data}
\name{spi.a}
\alias{spi.a}
\title{State personal income, annual.}
\format{Data frame with 1 row per state per year.
\describe{
\item{stabbr}{State postal abbreviation, factor}
\item{year}{Calendar year of data, numeric}
\item{value}{Personal income, $ thousands, numeric}
}}
\source{
Bureau of Economic Analysis, via API, using library(apitools)
}
\usage{
spi.a
}
\description{
State personal income, annual, 1929-2014. Includes 50 states, DC, and US.
}
\examples{
  spi.a
}
\keyword{datasets}

