% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spi.q.r
\docType{data}
\name{spi.q}
\alias{spi.q}
\title{State personal income, quarterly.}
\format{Data frame with 1 row per state and quarter.
\describe{
\item{stabbr}{State postal abbreviation, factor}
\item{date}{First day of quarter, date}
\item{value}{Personal income, $ thousands, numeric}
}}
\source{
Bureau of Economic Analysis, via API, using library(apitools)
}
\usage{
spi.q
}
\description{
State personal income, quarterly, 1948 through latest. Includes 50 states, DC, and US.
}
\examples{
  spi.q
}
\keyword{datasets}

