% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spovrates.r
\docType{data}
\name{spovrates}
\alias{spovrates}
\title{State poverty rates, annual.}
\format{Data frame with 1 row per state per year
\describe{
\item{stabbr}{State postal abbreviation, character}
\item{year}{Calendar year, numeric}
\item{povrate}{State poverty rate as a percentage, numeric}
}}
\source{
Bureau of the Census, prepared by Lucy Dadayan.
  Table 21. Number of Poor and Poverty Rate, by State [XLS – 125k]
  http://www.census.gov/hhes/www/poverty/data/historical/hstpov21.xls
See source code on GitHub.
}
\usage{
spovrates
}
\description{
State poverty rates, annual, 2000 through most recent, from the Census Bureau. One row per state per year.
\cr\cr
}
\details{
\cr\cr
Includes all 50 states, United States (US), and District of Columbia (DC) for all years.
}
\examples{
  spovrates
}
\keyword{datasets}

