% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recessions.r
\docType{data}
\name{recessions}
\alias{recessions}
\title{Recession dates.}
\format{Data frame with 1 row per recession
\describe{
\item{year}{Calendar year in which recession began}
\item{peak}{Date - first day of month in which recession began}
\item{trough}{Date - first day of month in which recession ended}
\item{rqsd}{Date - first day of quarter in which recession began}
\item{rqed}{Date - first day of quarter in which recession ended}
}}
\source{
National Bureau of Economic Research (NBER)
 http://www.nber.org/cycles.html
}
\usage{
recessions
}
\description{
Peak and trough dates for recessions.
}
\examples{
  recessions
}
\keyword{datasets}

